Extended the change hostname part by adding 3 new default variables and enhancing the change
Default Values for setting SAP Hostname according SAP Note#2002167

Normally ansible_hostname, etc. is used on the admin network, while the service hostname that
need to be checked is a different one.
In this case, or if DNS is not configured properly, you need to set the following variables explizitly
By introducing the variables, they can be overwritten, but the default is not changed

sap_preconfigure_hostname: '{{ ansible_hostname }}'
sap_preconfigure_domain: '{{ ansible_domain }}'
sap_preconfigure_ip: '{{ ansible_default_ipv4.address }}'

Added the required statements to enable this role in check mode